Module partiql_ast::ast

A PartiQL abstract syntax tree (AST).

This module contains the structures for the language AST. Two main entities in the module are Item and AstNode. AstNode represents an AST node and Item represents a PartiQL statement type, e.g. query, data definition language (DDL) data manipulation language (DML).

Structs

Assignment

Represents <expr> = <expr> in a DML SET operation. Note that in this case, = is representing an assignment operation and not the equality operator.

AstNode

Represents an AST node.

Bag

BagOpExpr

Between

BinOp

Call

CallAgg

CallArgNamed

CallArgNamedType

CreateIndex

CreateTable

CustomType

Ddl

Delete

Dml

DropIndex

DropTable

ExprPair

A generic pair of expressions. Used in the pub struct, searched_case and simple_case expr variants above.

FromClause

FROM clause of an SFW query

FromLet

GroupByExpr

GROUP BY <grouping_strategy> <group_key>[, <group_key>]… [AS ]

GroupKey

[AS ]

HavingClause

In

Insert

InsertValue

Join

Let

LetBinding

Like

LimitOffsetClause

List

NodeId

OnConflict

ON CONFLICT <expr> <conflict_action>

OrderByExpr

ORDER BY <sort_spec>…

Path

PathExpr

ProjectAll

ProjectExpr

ProjectPivot

Projection

Query

QueryTable

Remove

ReturningColumn

ReturningElem

<returning mapping> (<expr> [, <expr>]...)

ReturningExpr

RETURNING (<returning_elem> [, <returning_elem>]...)

SearchedCase

Select

Set

Sexp

SimpleCase

SortSpec

[ASC | DESC] ?

Struct

SymbolPrimitive

TopLevelQuery

UniOp

VarRef

WhereClause

WithClause

WithElement

Enums

BagOperator

BinOpKind

CallArg

Case

CaseSensitivity

Is used to determine if variable lookup should be case-sensitive or not.

ColumnComponent

ConflictAction

CONFLICT_ACTION <action>

CustomTypeParam

CustomTypePart

DdlOp

DmlOp

Expr

The expressions that can result in values.

FromLetKind

Indicates the type of FromLet, see the following for more details: https:///github.com/partiql/partiql-lang-kotlin/issues/242

FromSource

GroupingStrategy

Desired grouping qualifier: ALL or PARTIAL. Note: the group_ prefix is needed to avoid naming clashes.

Item

JoinKind

Indicates the logical type of join.

JoinSpec

Lit

Lit is mostly inspired by SQL-92 Literals standard and PartiQL specification. See section 5.3 in the following: https://www.contrib.andrew.cmu.edu/~shadow/sql/sql1992.txt and Section 2 of the following (Figure 1: BNF Grammar for PartiQL Values): https://partiql.org/assets/PartiQL-Specification.pdf

NullOrderingSpec

OrderingSpec

PathStep

A “step” within a path expression; that is the components of the expression following the root.

ProjectItem

An item to be projected in a SELECT-list.

ProjectionKind

Indicates the type of projection in a SFW query.

QuerySet

ReturningMapping

( MODIFIED | ALL ) ( NEW | OLD )

ScopeQualifier

Indicates scope search order when resolving variables. Has no effect except within FROM sources.

SetQuantifier

Indicates if a set should be reduced to its distinct elements or not.

Type

Represents all possible PartiQL data types.

UniOpKind

Type Aliases

AstTypeMap